refactor: use plugin system for aside
