+++ src/components/SlateEditor/helpers.ts
@@ -6,7 +6,7 @@
  *
  */
 import { ElementType } from "./interfaces";
-import { TYPE_ASIDE } from "./plugins/aside/types";
+import { ASIDE_ELEMENT_TYPE } from "./plugins/aside/asideTypes";
 import { AUDIO_ELEMENT_TYPE } from "./plugins/audio/audioTypes";
 import { TYPE_CAMPAIGN_BLOCK } from "./plugins/campaignBlock/types";
 import { TYPE_CODEBLOCK } from "./plugins/codeBlock/types";
@@ -41,7 +41,7 @@ export const inlines: ElementType[] = [
 ];
 
 export const blocks: ElementType[] = [
-  TYPE_ASIDE,
+  ASIDE_ELEMENT_TYPE,
   FRAMED_CONTENT_ELEMENT_TYPE,
   TYPE_CODEBLOCK,
   TYPE_DETAILS,
