--- src/components/SlateEditor/plugins/aside/index.tsx
@@ -1,83 +1,0 @@
-/**
- * Copyright (c) 2017-present, NDLA.
- *
- * This source code is licensed under the GPLv3 license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-import { Descendant, Editor, Element } from "slate";
-import { jsx as slatejsx } from "slate-hyperscript";
-import { TYPE_ASIDE } from "./types";
-import { createDataAttributes, createHtmlTag } from "../../../../util/embedTagHelpers";
-import { SlateSerializer } from "../../interfaces";
-import { defaultBlockNormalizer, NormalizerConfig } from "../../utils/defaultNormalizer";
-import {
-  afterOrBeforeTextBlockElement,
-  firstTextBlockElement,
-  lastTextBlockElement,
-  textBlockElements,
-} from "../../utils/normalizationHelpers";
-import { TYPE_PARAGRAPH } from "../paragraph/types";
-
-export interface AsideElement {
-  type: "aside";
-  data: {
-    type: "factAside";
-  };
-  children: Descendant[];
-}
-
-const normalizerConfig: NormalizerConfig = {
-  nodes: {
-    allowed: textBlockElements,
-    defaultType: TYPE_PARAGRAPH,
-  },
-  previous: {
-    allowed: afterOrBeforeTextBlockElement,
-    defaultType: TYPE_PARAGRAPH,
-  },
-  next: {
-    allowed: afterOrBeforeTextBlockElement,
-    defaultType: TYPE_PARAGRAPH,
-  },
-  firstNode: {
-    allowed: firstTextBlockElement,
-    defaultType: TYPE_PARAGRAPH,
-  },
-  lastNode: {
-    allowed: lastTextBlockElement,
-    defaultType: TYPE_PARAGRAPH,
-  },
-};
-
-export const asideSerializer: SlateSerializer = {
-  deserialize(el: HTMLElement, children: Descendant[]) {
-    if (el.tagName.toLowerCase() !== "aside") return;
-    return slatejsx("element", { type: TYPE_ASIDE, data: { type: "factAside" } }, children);
-  },
-  serialize(node, children) {
-    if (!Element.isElement(node)) return;
-    if (node.type !== "aside") return;
-    const data = createDataAttributes({ type: node.data.type });
-    return createHtmlTag({ tag: "aside", data, children });
-  },
-};
-
-export const asidePlugin = (editor: Editor) => {
-  const { normalizeNode: nextNormalizeNode } = editor;
-
-  editor.normalizeNode = (entry) => {
-    const [node, path] = entry;
-
-    if (Element.isElement(node) && node.type === TYPE_ASIDE) {
-      if (!defaultBlockNormalizer(editor, node, path, normalizerConfig)) {
-        return nextNormalizeNode(entry);
-      }
-    } else {
-      nextNormalizeNode(entry);
-    }
-  };
-
-  return editor;
-};
