+++ src/components/SlateEditor/plugins/aside/queries/asideQueries.ts
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) 2025-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import { isElementOfType } from "@ndla/editor";
+import { Node } from "slate";
+import { ASIDE_ELEMENT_TYPE } from "../asideTypes";
+
+export const isAsideElement = (node: Node | undefined) => isElementOfType(node, ASIDE_ELEMENT_TYPE);
