--- src/components/SlateEditor/plugins/aside/types.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright (c) 2022-present, NDLA.
- *
- * This source code is licensed under the GPLv3 license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-export const TYPE_ASIDE = "aside";
