+++ src/components/SlateEditor/plugins/aside/utils.ts
@@ -7,8 +7,8 @@
  */
 
 import { jsx as slatejsx } from "slate-hyperscript";
-import { TYPE_ASIDE } from "./types";
+import { ASIDE_ELEMENT_TYPE } from "./asideTypes";
 import { defaultParagraphBlock } from "../paragraph/utils";
 
 export const defaultAsideBlock = () =>
-  slatejsx("element", { type: TYPE_ASIDE, data: { type: "factAside" } }, defaultParagraphBlock());
+  slatejsx("element", { type: ASIDE_ELEMENT_TYPE, data: { type: "factAside" } }, defaultParagraphBlock());
