+++ src/components/SlateEditor/plugins/blockPicker/actions.tsx
@@ -34,7 +34,7 @@ import {
 import { DRAFT_ADMIN_SCOPE } from "../../../../constants";
 import HowToHelper from "../../../HowTo/HowToHelper";
 import { StoryType } from "../../../HowTo/stories";
-import { TYPE_ASIDE } from "../aside/types";
+import { ASIDE_ELEMENT_TYPE } from "../aside/asideTypes";
 import { AUDIO_ELEMENT_TYPE } from "../audio/audioTypes";
 import { TYPE_CAMPAIGN_BLOCK } from "../campaignBlock/types";
 import { TYPE_CODEBLOCK } from "../codeBlock/types";
@@ -72,7 +72,7 @@ export interface Action {
 
 export const commonActions: Action[] = [
   {
-    data: { type: TYPE_ASIDE, object: "factAside" },
+    data: { type: ASIDE_ELEMENT_TYPE, object: "factAside" },
     icon: <FileListFill />,
     helpIcon: renderArticleInModal("FactAside"),
   },
