+++ src/components/SlateEditor/utils/normalizationHelpers.ts
@@ -8,7 +8,7 @@
 import { Element } from "slate";
 import { jsx as slatejsx } from "slate-hyperscript";
 import { ElementType } from "../interfaces";
-import { TYPE_ASIDE } from "../plugins/aside/types";
+import { ASIDE_ELEMENT_TYPE } from "../plugins/aside/asideTypes";
 import { AUDIO_ELEMENT_TYPE } from "../plugins/audio/audioTypes";
 import { TYPE_QUOTE } from "../plugins/blockquote/types";
 import { TYPE_CODEBLOCK } from "../plugins/codeBlock/types";
@@ -46,7 +46,7 @@ export const textBlockElements: Element["type"][] = [
   TYPE_H5P,
   TYPE_FILE,
   TYPE_CODEBLOCK,
-  TYPE_ASIDE,
+  ASIDE_ELEMENT_TYPE,
   FRAMED_CONTENT_ELEMENT_TYPE,
   TYPE_DETAILS,
   TYPE_DEFINITION_LIST,
